DOMA-1852 move search to property domain
