+++ apps/condo/domains/ticket/utils/clientSchema/search.ts
@@ -38,16 +38,6 @@ const GET_ALL_PROPERTIES_BY_VALUE_QUERY = gql`
     }
 `
 
-const GET_ALL_PROPERTIES_WITH_MAP_BY_VALUE_QUERY = gql`
-    query selectProperty ($where: PropertyWhereInput, $orderBy: String, $first: Int, $skip: Int) {
-        objs: allProperties(where: $where, orderBy: $orderBy, first: $first, skip: $skip) {
-            id
-            address
-            map { sections { floors { units { label } } } }
-        }
-    }
-`
-
 const GET_ALL_DIVISIONS_BY_VALUE_QUERY = gql`
     query selectDivision ($where: DivisionWhereInput, $orderBy: String) {
         objs: allDivisions(where: $where, orderBy: $orderBy, first: 10) {
@@ -113,14 +103,6 @@ export async function searchProperty (client, where, orderBy, first = 10, skip =
     const { data = [], error } = await _search(client, GET_ALL_PROPERTIES_BY_VALUE_QUERY, { where, orderBy, first, skip })
     if (error) console.warn(error)
     if (data) return data.objs.map(x => ({ text: x.address, value: x.id }))
-
-    return []
-}
-
-export async function searchPropertyWithMap (client, where, orderBy, first = 10, skip = 0) {
-    const { data = [], error } = await _search(client, GET_ALL_PROPERTIES_WITH_MAP_BY_VALUE_QUERY, { where, orderBy, first, skip })
-    if (error) console.warn(error)
-    if (data) return data.objs.map(x => ({ address: x.address, id: x.id, map: x.map }))
 
     return []
 }
